Check auth server-side and centralize in game layout
Home page was a client component using useSession with a 1-second timeout to redirect unauthenticated users, causing a flash and potential mis-redirect on slow connections. Converted to a server component with an immediate getServerSession check; extracted the interactive form to HomeForm.

Added a layout for all game routes so auth is enforced in one place rather than per-page. Wrapped the Supabase user fetch in React.cache so layout and page calling getUser() in the same request only hits the DB once.
