+++ .gitignore
@@ -24,6 +24,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+log.txt
 
 # local env files
 .env*.local
